Add newsletter page (#41)
This adds a Newsletter page (located at /newsletter) accessible from a link in the Footer. This is primarily for SEO and so the newsletter page can be easily shared.

Previously the Newsletter form was in the Footer and it was global. Because it felt redundant I've since removed it and moved it to the Blog post page template.

Also:

- Change color of Blog page title
- Reorg some files
- Change font-size of navItems
- Update Footer/Work page styles